Add files via upload
Welcome to this project on the Numpy - Cat vs Non-cat Classifier with Logistic Regression using Numpy. In this project, you will use Python and Numpy to build a Logistic Regression Classifier from scratch, and apply it to predict the class of an input image - whether it is a cat or a non-cat.

Though we have a lot of ready-made APIs like scikit-learn and Keras to build Machine Learning and Deep Learning models, it is very essential for a Machine Learning enthusiast to clearly understand the hidden mechanism behind the working of ML models. Upon completing this project, you will understand the pragmatic application of various Numpy functions with a clear picture of workflow in an end-to-end Machine Learning project.

Skills you will develop:

Numpy
Python Programming
H5py files
Machine Learning