Add files via upload
Welcome to this project on Classifying Flowers in Iris dataset with Deep Neural Network using Keras. In this project, you will use Python and Keras to build a Deep Neural Network, and apply it to predict the classes of Flowers in the Iris dataset.

Keras is one of the most extensively used APIs in the world of Deep Learning. It provides an amazing developer-friendly deep learning framework to build deep learning models with wide-ranging features to support high scalability, because of which it is not only widely used in academics but also in organizations to build state-of-the-art research models. In this exercise, you will get a step-wise hands-on guide on how to build a basic deep neural network using Keras, and how to use it to predict the class of the test samples.

Skills you will develop:

Keras
Deep Learning
Python
Pandas
Matplotlib