Add files via upload
Welcome to this project on Cat vs Non-cat Classifier using Transfer Learning. In this project, you will use Python and Keras to apply the Transfer Learning technique in order to build an image classifier, and apply it to predict the class of an input image - whether it is a cat or a non-cat.

Deep Learning is computationally intensive, often demanding powerful computational resources to yield reasonable accuracies in the real world. The idea of Transfer Learning has become a boon for the computer vision and deep learning community as it has reduced the hunger of Deep Learning algorithms for powerful hardware resources. In addition, Transfer Learning has made it possible to yield impressive results within a short amount of time, even with fewer data. Upon completing this project, you will understand the pragmatic application of the transfer learning technique using Keras, and you will be able to appreciate its salient features.

Skills you will develop:

Transfer Learning
Keras
Python Programming
Tensorflow