Add files via upload
Credit Card Fraud Detection
Welcome to this project on Credit Card Fraud Detection. In this project, you will use Python, SMOTE Technique(to over-sample data), build a Logistic Regression Classifier, and apply it to detect if a transaction is fraudulent or not.

The real world datasets often might be with data of imbalanced classes. It is very important to feed a decent number of data samples of each class in a classification problem so that the classifier would detect the underlying hidden patterns for each class and prepare itself to reasonably classify the test data. Upon completing this project, you will understand the pragmatic application of various Pandas functions, with a clear picture of how to over-sample the dataset with imbalanced classes using the SMOTE technique and how to use the thus obtained data to train a classifier.

Skills you will develop:

Pandas
Python Programming
SMOTE
Scikit-Learn